SFML WIP Update 6
Cleaned up code, Added Documentation
Finishing up CluwneLib, Beginning to work on GameStates

+++ SS14.Client.Graphics.CluwneLib/Viewport.cs
@@ -7,8 +7,8 @@ namespace SS14.Client.Graphics.CluwneLib
 {
     public class Viewport
     {
-        public float Width { get; set; }
-        public float Height { get; set; }
+        public int Width { get; set; }
+        public int Height { get; set; }
 
     }
 }
